% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uts.R
\name{as.data.frame.uts}
\alias{as.data.frame.uts}
\title{Coerce to a Data Frame}
\usage{
\method{as.data.frame}{uts}(x, ...)
}
\arguments{
\item{x}{a \code{"uts"} object.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Flatten a \code{"uts"} to a two-column \code{data.frame}, with the observation times in first column and observation values in the second column. The column names are \code{"time"} and \code{"value"}.
}
\note{
Only time series with atomic observation values can be coerced to a \code{data.frame}.

This method is helpful for saving a time series to a human-readable text file.
}
\examples{
as.data.frame(ex_uts())

# Save a time series to a text file with nice date formatting
uts1 <- uts(rnorm(10), Sys.time() + days(1:10))
data <- as.data.frame(uts1)
data$time <- format(data$time, format="\%Y-\%m-\%d")
\dontrun{write.csv(data, file="random.csv", row.names=FALSE, quote=FALSE)}
}
