% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercions.R
\name{as.irts.uts}
\alias{as.irts.uts}
\title{Coercion to irts}
\usage{
as.irts.uts(x)
}
\arguments{
\item{x}{a \code{"uts"} object.}
}
\value{
An \code{\link[tseries]{irts}} object.
}
\description{
Coercion to irts
}
\examples{
if (requireNamespace("tseries", quietly = TRUE)) {
  tseries::as.irts(ex_uts())
}
}

