% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercions.R
\name{as.its.uts}
\alias{as.its.uts}
\title{Coercion to its}
\usage{
as.its.uts(x, ...)
}
\arguments{
\item{x}{a \code{"uts"} object.}

\item{\dots}{further arguments passed to or from methods.}
}
\value{
An \code{\link[its]{its}} object.
}
\description{
Coercion to its
}
\examples{
if (requireNamespace("its", quietly = TRUE)) {
  its::as.its(ex_uts())
}
}
