% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/POSIXt_extensions.R
\name{as.POSIXct_matrix}
\alias{as.POSIXct_matrix}
\title{Coerce to POSIXct_matrix}
\usage{
as.POSIXct_matrix(x, ...)
}
\arguments{
\item{x}{an \R{} object of appropriate type.}

\item{\dots}{arguments passed to or from methods.}
}
\description{
Generic function to coerce an object to a POSIXct_matrix.
}
\details{
See documentation on method functions for further details.
}
\seealso{
\code{\link{POSIXct_matrix}}
}

