% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercions.R
\name{as.xts.uts}
\alias{as.xts.uts}
\title{Coercion to xts}
\usage{
as.xts.uts(x, ...)
}
\arguments{
\item{x}{a \code{"uts"} object with atomic observation values.}

\item{\dots}{further arguments passed to or from methods.}
}
\value{
An \code{\link[xts]{xts}} object.
}
\description{
Coercion to xts
}
\examples{
if (requireNamespace("xts", quietly = TRUE)) {
  xts::as.xts(ex_uts())
}
}

