% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercions.R
\name{as.zoo.uts}
\alias{as.zoo.uts}
\title{Coercion to zoo}
\usage{
as.zoo.uts(x, ...)
}
\arguments{
\item{x}{a \code{"uts"} object.}

\item{\dots}{further arguments passed to or from methods.}
}
\value{
A \code{\link[zoo]{zoo}} object.
}
\description{
Coercion to zoo
}
\examples{
if (requireNamespace("zoo", quietly = TRUE)) {
  zoo::as.zoo(ex_uts())
}
}

