% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_methods.R
\name{binary_Ops}
\alias{binary_Ops}
\title{Binary Ops Groups Methods}
\usage{
binary_Ops(x, Ops, y, times = "all", interpolation = "last")
}
\arguments{
\item{x, y}{either \code{\link{uts}} objects or compatible length-one \R objects, where compatability depends on the type of operation performed.}

\item{Ops}{a binary \code{\link{Ops}} operator.}

\item{times}{which observation times to use for the output time series: \itemize{
   \item \code{"all"}: the union of observation times of \code{x} and \code{y}.
   \item \code{"x"}: the observation times of \code{x}.
   \item \code{"y"}: the observation times of \code{y}.
}
In all three cases, times before the first observation time (i.e. the start time) of either time series are excluded.
The method for determining these times is unaffected by numerical noise less than \code{sqrt(\link[=.Machine]{.Machine$double.eps})}.}

\item{interpolation}{}
}
\description{
Apply a binary \code{\link{Ops}} group method to the observation values of \code{"uts"} objects.
}
\examples{

# Create two sample time series
x <- ex_uts()
y <-  head(x * 1.1, dhours(1)), 5)

# Vary interpolation method
binary_Ops(x, "/", y)
binary_Ops(x, "/", y, interpolation="linear")

# Vary time point in output
binary_Ops(x, "/", y)
binary_Ops(x, "/", y, times="x")
binary_Ops(x, "/", y, times="y")
}
\seealso{
\code{\link{Ops.uts}}, \code{\link{groupGeneric}}
}

