% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diff.R
\name{diff_t.uts}
\alias{diff_t.uts}
\title{Rolling Differences}
\usage{
\method{diff_t}{uts}(x, ...)
}
\arguments{
\item{x}{a \code{"uts"} object.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Rolling Differences
}
\examples{
diff_t(ex_uts())
}
\seealso{
\code{\link{lag}}, \code{\link{lag_t}}, and \code{\link{diff}}
}

