% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/example_ts.R
\name{download_co2}
\alias{download_co2}
\title{Mauna Loa Atmospheric CO2 Concentration}
\usage{

  download_co2(file = "ftp://aftp.cmdl.noaa.gov/products/trends/co2/co2_mm_mlo.txt")
}
\arguments{
\item{file}{the download location of the CO2 dataset.}
}
\description{
This function downloads the monthly average atmospheric CO2 concentration (in parts per million), as measured at the Mauna Loa observatory, from the NOAA website. Please see \url{http://www.esrl.noaa.gov/gmd/ccgg/trends/} for a detailed description. The downloaded data is subsequently imported into \R and returned as a \code{\link{uts}} object.
}
\examples{
co2_ml <- download_co2()
plot(co2_ml)
table(diff(time(co2_ml)))    # most observations are one month apart
}
\seealso{
The \code{\link[datasets:co2]{co2}} dataset in base \R is very similar, but ends in 1997 and has several missing values filled in using linear interpolation.
}
\keyword{datasets}

