% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/example_ts.R
\name{ex_uts}
\alias{ex_uts}
\alias{ex_uts2}
\title{Example Time Series}
\usage{
ex_uts()

ex_uts2()
}
\value{
\code{ex_ust()} returns a numeric \code{"uts"} with six observations.

\code{ex_uts2()} returns a non-numeric \code{"uts"}  with three observations.
}
\description{
Create time series that can be used for code examples and testing.
}
\examples{
ex_uts()
ex_uts2()
}

