% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/example_ts.R
\name{example_ts}
\alias{ex_uts}
\alias{ex_uts2}
\alias{example_ts}
\title{Example Time Series}
\usage{
ex_uts()

ex_uts2()
}
\description{
Create time series that can be used for code examples and testing.
}
\section{Functions}{
\itemize{
\item \code{ex_uts}: an a numeric \code{\link{uts}} with six observations.

\item \code{ex_uts2}: a non-numeric \code{\link{uts}} with three observations
}}
\examples{
ex_uts()
ex_uts2()
}

