% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/operators.R
\name{group_generic_uts}
\alias{group_generic_uts}
\title{S3 Group Generic Methods for uts}
\arguments{
\item{x}{an object of class \code{"uts"}.}

\item{\dots}{further arguments passed to or from methods.}
}
\value{
An object of class \code{"uts"} with same observation times as the input \code{x}, and observation values the result of applying the provided method to the observation values of input \code{x}.
}
\description{
S3 group generic methods apply the methods in base \R{} to the values of \code{\link{uts}} objects.
}
\seealso{
\code{\link{groupGeneric}}, \code{\link{Math}}, \code{\link{Ops}}, \code{\link{Summary}}
}

