% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/head_tail.R
\name{head.uts}
\alias{head.uts}
\alias{tail.uts}
\title{Return the First or Last Part}
\usage{
\method{head}{uts}(x, n = 6L, ...)

\method{tail}{uts}(x, n = 6L, ...)
}
\arguments{
\item{x}{a \code{"uts"} object.}

\item{n}{a single integer. If positive, length of the the resulting time series. If negative, all but the \code{n} last/first number of observations of \code{x}.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Return a subperiod time series with the first or last part.
}
\examples{
head(ex_uts(), 2)
head(ex_uts(), -2)
head(ex_uts(), -6)
tail(ex_uts(), 2)
tail(ex_uts(), -2)
tail(ex_uts(), -6)
}
\seealso{
\code{\link{head}} and \code{\link{tail}} in base \R.
}
\keyword{internal}

