% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_t_tail_t.R
\name{head_t}
\alias{head_t}
\alias{head_t.uts}
\title{Initial Subperiod}
\usage{
head_t(x, ...)

\method{head_t}{uts}(x, width, ...)
}
\arguments{
\item{x}{a time series object.}

\item{width}{a \code{\link[lubridate]{duration}} object, specifying the temporal width of the initial subperiod time series.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Return a subperiod time series with the initial part, where the length is specified in terms of the \emph{temporal length}.
}
\section{Methods (by class)}{
\itemize{
\item \code{uts}: initial subperiod time seried for \code{"uts"} object.
}}
\note{
For an evenly spaced time series, the methods \code{head()} and \code{head_t()} essentially work the same. For unevenly spaced time series, however, they are quite different. The subperiod window of the former is determined by the observation \emph{values}, while for the latter it is determined by the observation \emph{times}.
}
\examples{
head_t(ex_uts(), ddays(1))
head_t(ex_uts(), ddays(0))  # leaves only the first observation
}
\seealso{
\code{\link{head}}, \code{\link{tail}}, \code{\link{tail_t}}, \code{\link{window}} for other methods that extract a subperiod time series.
}

