% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaN_finite.R
\name{is.finite.uts}
\alias{is.finite.uts}
\alias{is.infinite.uts}
\alias{is.nan.uts}
\title{Finite, Infinite and NaN Numbers}
\usage{
\method{is.finite}{uts}(x)

\method{is.infinite}{uts}(x)

\method{is.nan}{uts}(x)
}
\arguments{
\item{x}{a \code{"uts"} object.}
}
\value{
A logical \code{"uts"} (i.e. a \code{"uts"} with \code{\link{logical}} observation values), indicating which observation values are finite, infinite, or \code{NaN}, respectively.
}
\description{
Find finite, infinite, and NaN observation values.
}
\examples{
# Create sample time series
test <- ex_uts()
test$values[c(2, 4)] <- c(NaN, -Inf)

# Get logical "uts", indicating which observations are finite, infinte, or NaN
is.finite(test)
is.infinite(test)
is.nan(test)
}
\seealso{
\code{\link[base:is.finite]{is.finite}}, \code{\link[base:is.infinite]{is.infinite}}, \code{\link[base:is.nan]{is.nan}} in base \R.
}

