% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NAs.R
\name{is.na.uts}
\alias{is.na.uts}
\title{Not Available / Missing Values}
\usage{
\method{is.na}{uts}(x)
}
\arguments{
\item{x}{a \code{"uts"} object.}
}
\description{
Find missing values.
}
\details{
\code{is.na} returns a logical \code{"uts"}, i.e. a \code{"uts"} with \code{\link{logical}} observation values, indication which observation values are \code{NA}.
}
\examples{
# Set observation to NA
test <- ex_uts()
test$values[c(2, 4)] <- NA      # NEXT: use replacement using time points

# Get logical "uts", indicating which observations are NA
is.na(test)
}
\seealso{
\code{\link{na.omit}}
}

