% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/POSIXct_matrix.R
\name{is.POSIXct_matrix}
\alias{is.POSIXct_matrix}
\title{Check if Object is POSIXct_matrix}
\usage{
is.POSIXct_matrix(x)
}
\arguments{
\item{x}{an \R object.}
}
\description{
Return \code{TRUE} if \code{x} is an object of class \code{"POSIXct_matrix"}.
}

