% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/POSIXct_vector.R
\name{is.POSIXct_vector}
\alias{is.POSIXct_vector}
\title{Check if Object is POSIXct_vector}
\usage{
is.POSIXct_vector(x)
}
\arguments{
\item{x}{an \R object.}
}
\description{
Return \code{TRUE} if \code{x} is an object of class \code{"POSIXct_vector"}.
}

