% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uts.R
\name{is.uts}
\alias{is.uts}
\title{Is Object a uts?}
\usage{
is.uts(x)
}
\arguments{
\item{x}{an \R object.}
}
\description{
Return \code{TRUE} if and only if the argument is a \code{"uts"} object.
}
\examples{
is.uts(ex_uts())
is.uts(5)
}
\keyword{internal}

