% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uts.R
\name{is.uts_virtual}
\alias{is.uts_virtual}
\title{Is Object a uts_virtual?}
\usage{
is.uts_virtual(x)
}
\arguments{
\item{x}{an \R object.}
}
\description{
Return \code{TRUE} if and only if the argument is a \code{"uts_virtual"} object.
}
\examples{
is.uts_virtual(ex_uts())
is.uts_virtual(5)
}
\keyword{internal}

