% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uts.R
\name{lag.uts}
\alias{lag.uts}
\title{Lag a Time Series}
\usage{
\method{lag}{uts}(x, k = 1, ...)
}
\arguments{
\item{x}{a time series object.}

\item{k}{the number of lags (in units of observations).}

\item{\dots}{further arguments passed to or from methods.}
}
\value{
A time series object with the same class, length, and observation times as \code{x}.
}
\description{
Compute a lagged version of a time series by shifting individual observations values, while keeping the observation times unchanged.
}
\details{
The n-th observation of each original time series becomes the (n+k)-th observation of the lagged time series for 1 <= (n+k) <= length(x). Observations without corresponding un-lagged value (for example, the second observation for lag k=3) are set to \code{NA}.
}
\examples{
# Shift observations values forward by one observation
lag(ex_uts(), k=1)

# Shift observations values forward by two observations
lag(ex_uts(), k=-2)

# If the lag >= the length of the time series, all observation values are N
lag(ex_uts(), k=6)
}
\seealso{
\code{\link[stats:lag]{lag}}
}

