% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lag.R
\name{lag_t}
\alias{lag_t}
\alias{lag_t.uts}
\title{Lag Observation Times}
\usage{
lag_t(x, ...)

\method{lag_t}{uts}(x, lag_t, ...)
}
\arguments{
\item{x}{a time series object of appropriate type.}

\item{lag_t}{a \code{\link[lubridate]{duration}} object, specifying how much to shift the observation times of \code{x} forward.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Lag observation times of a time series by a given amount. In other words, add a certain amount of time to each observation time.
}
\section{Methods (by class)}{
\itemize{
\item \code{uts}: shift observation times for a \code{"uts"} object.
}}
\note{
For an evenly spaced time series (1) shifting observation \emph{times}, and (2) shifting observation \emph{values} essentinally gives the same result. For unevenly spaced time series, however, these two operations are quite different. The former affects only the observation \emph{times}, while the latter affects only the observation \emph{values} (apart from observations that are dropped).
}
\examples{
lag_t(ex_uts(), ddays(3))
lag_t(ex_uts(), dhours(7.5))
}
\seealso{
\code{\link{lag.uts}} allows to shift observation \emph{values}, as opposed to observation \emph{times}.
}

