% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_last.R
\name{last}
\alias{last}
\alias{last.default}
\alias{last.uts}
\title{Last Observation Value}
\usage{
last(x, ...)

\method{last}{default}(x, ...)

\method{last}{uts}(x, ...)
}
\arguments{
\item{x}{an \R object.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Return the last value of an \R object.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Assume that \code{x} is a vector or compatible object and return the last value.

\item \code{uts}: Return the last observation \emph{value} of a \code{"uts"} object. Unlike method \code{last()} in package \href{https://cran.r-project.org/web/packages/xts/}{xts}, arguments \code{n} and \code{keep} are not supported. Instead, this functionality is provided by \code{tail()} and \code{tail_t()}.
}}
\examples{
last(5:10)
last(list(a="test", b=3))
first(ex_uts())
}
\seealso{
\code{\link{first}}

\code{\link{start}} and \code{\link{end}} return, respectively, the first and last obvservation \emph{time} from a time series.
}

