% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uts.R
\name{length.uts}
\alias{length.uts}
\title{Length of an Object}
\usage{
\method{length}{uts}(x)
}
\arguments{
\item{x}{a time series object.}
}
\description{
For a univariate time series, get the number of observations. For a multi-variate time series, get the total number of time series.
}
\examples{
length(ex_uts())
}
\seealso{
\code{\link{length}}
}

