% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{length_t}
\alias{length_t}
\alias{length_t.uts}
\title{Temporal Length}
\usage{
length_t(x, ...)

\method{length_t}{uts}(x, ...)
}
\arguments{
\item{x}{a time series object.}

\item{\dots}{further arguments passed to or from methods.}
}
\value{
A \code{\link[lubridate]{duration}} object.
}
\description{
Return the temporal length of a time series, i.e. the time difference between the first and last observation time.
}
\section{Methods (by class)}{
\itemize{
\item \code{uts}: temporal length of \code{"uts"} object.
}}

\note{
For an evenly spaced time series there is a one-to-one correspondence between (1) the \emph{number} of observations, and (2) \emph{temporal} length. For unevenly spaced time series, however, these two concepts are quite different.
}
\examples{
length_t(ex_uts())
length_t(ex_uts2())
length_t(uts())
}
\seealso{
\code{\link{length.uts}} allows to get the \emph{number} observations.
}
