% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uts.R
\name{merge.uts}
\alias{merge.uts}
\title{Merge two or more uts}
\usage{
\method{merge}{uts}(x, y, tolerance = .Machine$double.eps^0.5, ...)
}
\arguments{
\item{x,y}{\code{"uts"} objects.}

\item{tolerance}{tolerance for numerical noise in observation times.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Use data of first-supplied times series, if there are multiple observations at a point in time
}
\examples{
merge(ex_uts(), ex_uts())
merge(ex_uts(), ex_uts2())
merge(uts(), ex_uts())
}
\seealso{
\code{\link{which}}
}

