% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NAs.R
\name{na.omit.uts}
\alias{na.omit.uts}
\title{Remove NAs}
\usage{
\method{na.omit}{uts}(object, ...)
}
\arguments{
\item{object}{a time series object.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Returns the object with incomplete cases removed.
}
\examples{
# Remove NAs from a "uts"
test <- ex_uts()
test$values[c(2, 4)] <- NA
na.omit(test)
}
\seealso{
\code{\link[=is.na.uts]{is.na}}

\code{\link[base:is.na]{is.na}}, \code{\link[stats:na.fail]{na.omit}} in base \R.
}

