% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/POSIXt_extensions.R
\name{next_business_day.POSIXt}
\alias{next_business_day.POSIXt}
\title{Next Business Day}
\usage{
\method{next_business_day}{POSIXt}(x)
}
\arguments{
\item{x}{a \code{POSIXt} date object.}
}
\description{
For a POSIXt date, get the next business day on or after such date. The time-of-day information is unaffected. In particular, if the input already is a business day, then the output is the same as the input.
}
\details{
This function is needed, because \code{\link[lubridate:ceiling_date]{ceiling_date()}} does not have support for unit="quarter".
}
\seealso{
\code{\link[lubridate:ceiling_date]{ceiling_date()}}
}

