% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/POSIXt_extensions.R
\name{next_business_day}
\alias{next_business_day}
\title{Next Business Day.}
\usage{
next_business_day(x, ...)
}
\arguments{
\item{x}{an object containing dates.}

\item{\dots}{arguments passed to or from methods.}
}
\description{
Generic function to find the next business day of objects representing dates.
}
\details{
See documentation on method functions for further details.
}
\seealso{
\code{\link{previous_business_day}}
}

