% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/POSIXt_extensions.R
\name{next_weekday}
\alias{next_weekday}
\alias{next_weekday.POSIXt}
\title{Next Weekday}
\usage{
next_weekday(x, ...)

\method{next_weekday}{POSIXt}(x, ...)
}
\arguments{
\item{x}{an object representing dates.}

\item{\dots}{further arguments passed to or from methods.}
}
\value{
An object of the same class and length as the input \code{x}
}
\description{
For objects representing dates, find the next weekday.
}
\details{
Add the minimum number of days to reach a weekday (Monday-Friday). The time-of-day information is therefore unaffected. In particular, if the input already is a weekday, then the output is the same as the input.
}
\section{Methods (by class)}{
\itemize{
\item \code{POSIXt}: find the next weekday for POSIXt dates
}}
\examples{
# 2015-04-20 is a Monday
next_weekday(as.POSIXct("2015-04-20") + days(0:6))
}
\seealso{
\code{\link{previous_weekday}}
}
\keyword{chron}

