% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_interfaces.R
\name{num_leq_sorted_R}
\alias{num_leq_sorted_R}
\title{R implementation of num_leq_sorted}
\usage{
num_leq_sorted_R(a, b)
}
\arguments{
\item{a}{a strictly increasing vector of numbers.}

\item{b}{a strictly increasing vector of numbers.}
}
\value{
An integer vector of same length as \code{a}.
}
\description{
This function is identical to \code{\link{num_leq_sorted}} except that a) the input vectors need to be strictly increasing, and b) there is no numerical noise tolerance support. It exists solely for testing the C implementation.
}
\examples{
num_leq_sorted_R(c(-3, 1, 3, 5, 7), c(0, 1, 4, 9, 16))
}
\keyword{internal}

