% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_interfaces.R
\name{num_less_sorted}
\alias{num_less_sorted}
\title{Less-Than Comparison of Sorted Vectors}
\usage{
num_less_sorted(a, b, tolerance = 0)
}
\arguments{
\item{a}{a sorted, i.e. non-decreasing, vector of numbers.}

\item{b}{a sorted, i.e. non-decreasing, vector of numbers.}
}
\value{
An integer vector of same length as \code{a}.
}
\description{
For two sorted numeric vectors \code{a} and \code{b}, for each element \code{a[i]} determine the number of elements in \code{b} that are less than this value.
}
\note{
Equivalently, because the input vectors are sorted, for each element \code{a[i]} determine the maximum index \code{j} with \code{b[j] < a[i]}.
}
\examples{
# The second vector has
# -) 0 elements less than -3
# -) 1 element less than 1
# -) 2 elements less than 3
# -) 3 elements less than 5
# -) 3 elements less than 7
num_less_sorted(c(-3, 1, 3, 5, 7), c(0, 1, 4, 9, 16))

# Trivial cases
num_less_sorted(1:5, 1:5)
num_less_sorted(c(), 1:5)
num_less_sorted(1:5, c())
}
\seealso{
\code{\link{num_leq_sorted}} for a less-than-or-equqal comparison of two sorted vectors.
}
\keyword{internal}

