% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onepopulation.R
\name{One_Population}
\alias{One_Population}
\alias{onepopulation}
\title{Inferences about Mean Vector}
\usage{
onepopulation(xbar, s, n, mu = 0, a = 0.05, sigma = c("known", "unknown"))
}
\arguments{
\item{xbar}{xbare}

\item{s}{matriks cova}

\item{n}{jumlah sampel}
}
\description{
Inferences about Mean Vector
}
\author{
Muhammad Sirojuddin
}
