% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manova.R
\name{OnewayManova}
\alias{OnewayManova}
\alias{oneway}
\title{One way MANOVA}
\usage{
oneway(xbar, s, n, a = 0.05, method = c("kecil", "besar"))
}
\arguments{
\item{xbar}{dadekno siji px2}

\item{s}{dadekno list(s1,s2)}

\item{n}{c(n1,n2)}

\item{method}{c("kecil","besar") ukuran sample
ci[\link{1}] menunjukkan ci antara grup 1 dan 2}
}
\description{
One way MANOVA
}
\author{
Muhammad Sirojuddin
}
