% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_methods.R
\name{Ops.list}
\alias{Ops.list}
\alias{Ops.uts}
\title{Ops Group Methods}
\usage{
\method{Ops}{list}(e1, e2)
}
\arguments{
\item{e1, e2}{either \code{"uts"} objects or numeric numbers.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Apply the \code{\link{Ops}} methods in base \R{} to \code{"uts"} objects.
}
\examples{
# Unary oparators
-ex_uts()
!ex_uts()

# Binary operators
ex_uts() * 2
2 * ex_uts()
ex_uts() / ex_uts()
ex_uts() > 48
48 >= ex_uts()
}
\seealso{
\code{\link{groupGeneric}}
}

