% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/group_methods.R
\name{Ops.uts}
\alias{Ops.uts}
\title{Ops Group Methods for uts}
\usage{
\method{Ops}{uts}(e1, e2)
}
\arguments{
\item{e1,e2}{either \code{"uts"} objects or numeric numbers.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Apply the \code{\link{Ops}} methods in base \R{} to the observation values of \code{"uts"} objects.
}
\note{
For unary oparations, the output time series has the same observation times as the input time series.

For binary operations involving two time series \code{uts1} and \code{uts2}, the output time series has the union of observation times of \code{uts1} and \code{uts2}, but excluding times before \emph{both} time series have their initial observation. The method for determining these times is unaffected by numerical noise less than \code{sqrt(\link[=.Machine]{.Machine$double.eps})}.
}
\examples{
# Unary oparators
-ex_uts()
!ex_uts()

# Binary operators
ex_uts() * 2
2 * ex_uts()
ex_uts() / ex_uts()
ex_uts() > 48
48 >= ex_uts()
}
\seealso{
\code{\link{groupGeneric}}
}

