% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_methods.R
\name{Ops_uts}
\alias{Ops_uts}
\title{Ops Group Methods Helper}
\usage{
Ops_uts(e1, e2, .Generic)
}
\arguments{
\item{e1, e2}{see \code{\link{Ops.uts}}.}

\item{.Generic}{a character vector naming the generic function.}
}
\description{
A helper function that does the actual work for the \code{\link{Ops.uts}}.
}
\examples{
Ops_uts(ex_uts(), ex_uts(), "+")         # arithmetic
Ops_uts(ex_uts(), ex_uts(), ">")         # comparison
!Ops_uts(ex_uts() > 48, .Generic="!")    # logical operator
}
\seealso{
\code{\link{groupGeneric}}
}
\keyword{internal}
