% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.uts}
\alias{plot.uts}
\title{Plot a uts}
\usage{
\method{plot}{uts}(x, type = "l", col = "blue", xlab = "", ylab = "",
  ...)
}
\arguments{
\item{x}{a numeric \code{\link{uts}} that should be plotted.}

\item{type}{what type of plot should be drawn, see \code{\link[graphics:plot]{plot}}. \code{type="b"} is helpful for highlighting individual observations.}

\item{col}{the colors for lines and points. See \code{\link[graphics:plot]{plot}}.}

\item{xlab}{a label for the x axis.}

\item{ylab}{a label for the y axis.}

\item{\dots}{other graphical parameters passed to \code{\link[graphics:plot]{plot}}.}
}
\description{
This function is convenience wrapper around \code{\link[graphics:plot.default]{plot.default}} that picks several sensible default arguments
}
\examples{
plot(ex_uts())

# Use custom date formatting for x axis
# -) this example produces several harmless warning messages, because only axis.POSIXct()
#    supports this argument
plot(ex_uts(), format="\%a \%I\%p")
}

