% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/POSIXct_matrix.R
\name{POSIXct_matrix.POSIXct_vector}
\alias{POSIXct_matrix.POSIXct_vector}
\title{Matrix of POSIXct Objects}
\usage{
\method{POSIXct_matrix}{POSIXct_vector}(x, nrow = 1, ncol = 1,
  byrow = FALSE, dimnames = NULL, ...)
}
\arguments{
\item{x}{coming soon}

\item{nrow}{coming soon}

\item{ncol}{coming soon}

\item{byrow}{coming soon}

\item{dimnames}{coming soon}

\item{\dots}{arguments passed to or from methods.}
}
\description{
Matrix of POSIXct Objects
}
\keyword{chron}
\keyword{classes}

