% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/POSIXct_vector.R
\name{POSIXct_vector.default}
\alias{POSIXct_vector.default}
\title{Vector of POSIXct Objects}
\usage{
\method{POSIXct_vector}{default}(...)
}
\arguments{
\item{\dots}{arguments passed to or from methods.}
}
\description{
Vector of POSIXct Objects
}
\keyword{chron}
\keyword{classes}

