% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/POSIXt_extensions.R
\name{POSIXct_vector}
\alias{POSIXct_vector}
\title{Vector of POSIXct Objects}
\usage{
POSIXct_vector(x, ...)
}
\arguments{
\item{x}{an \R{} object of appropriate type.}

\item{\dots}{arguments passed to or from methods.}
}
\description{
Generic function to create a vector of POSIXct Objects.
}
\details{
See documentation on method functions for further details.
}
\seealso{
\code{\link{POSIXct_matrix}}
}

