% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/POSIXt_extensions.R
\name{previous_business_day}
\alias{previous_business_day}
\title{Previous Business Day}
\usage{
previous_business_day(x, ...)
}
\arguments{
\item{x}{an object representing dates.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Generic function to find the previous business day of objects representing dates.
}
\details{
See documentation on method functions for further details.
}
\seealso{
\code{\link{next_business_day}}
}

