% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uts.R
\name{print.uts}
\alias{print.uts}
\title{Print Values}
\usage{
\method{print}{uts}(x, style = "horizontal", ...)
}
\arguments{
\item{x}{a time series object.}

\item{style}{the printing style. Either \code{"horizontal"} (the default), "vertical" or "plain" (which first prints the data and then the index).}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Print Values
}
\examples{
print(ex_uts())
print(ex_uts(), style="vertical")
print(ex_uts(), style="plain")
uts()
}
\seealso{
\code{\link[base:print]{print}}
}

