% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/POSIXt_extensions.R
\name{quarterend.POSIXt}
\alias{quarterend.POSIXt}
\title{Quarter End}
\usage{
\method{quarterend}{POSIXt}(x)
}
\arguments{
\item{x}{a \code{POSIXt} date object.}
}
\description{
For a POSIXt date, return the last day of the corresponding quarter.
}
\details{
This function is needed, because \code{\link[lubridate:ceiling_date]{ceiling_date()}} does not have support for unit="quarter".
}
\seealso{
\code{\link[lubridate:ceiling_date]{ceiling_date()}}
}

