% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/POSIXt_extensions.R
\name{quarterend}
\alias{quarterend}
\title{Generic Function to determine the Quarter-End of dates.}
\usage{
quarterend(x, ...)
}
\arguments{
\item{x}{an object containing dates.}

\item{\dots}{additional parameters. See documentation on individual methods.}
}
\description{
See documentation on method functions for further details.
}
\details{
This function is needed, because \code{\link[lubridate:ceiling_date]{ceiling_date()}} does not have support for unit="quarter".
}
\seealso{
\code{\link[lubridate:ceiling_date]{ceiling_date()}}
}

