% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoindependent.R
\name{Repeated}
\alias{Repeated}
\alias{repeated}
\title{Repeadted Measurements}
\usage{
repeated(xbar, s, n, c, a = 0.05)
}
\arguments{
\item{xbar}{dadekno siji px2}

\item{s}{dadekno list(s1,s2)}

\item{n}{c(n1,n2)}
}
\description{
Repeadted Measurements
}
\author{
Muhammad Sirojuddin
}
