% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_values}
\alias{sample_values}
\alias{sample_values.uts}
\title{Sample Values}
\usage{
sample_values(x, ...)

\method{sample_values}{uts}(x, sampling_times, method = "last",
  max_dt = ddays(Inf), tolerance = .Machine$double.eps^0.5, ...)
}
\arguments{
\item{x}{a time series object.}

\item{sampling_times}{a strictly increasing sequence of \code{\link{POSIXct}} date-times.}

\item{method}{the sampling method.
\code{"last"} returns the most recent observation value in \code{x} for each sampling time.
\code{"linear"} returns for each sampling time the linearily-interpolated  observation values in \code{x} that immediately preceede and follow each sampling time.}

\item{max_dt}{a duration object, specifying the maximum time difference between each sampling time and the observation times in \code{x} used for determining the sampled value. Return \code{NA} as sampled value for sampling times for which this threshold is exceeded.}

\item{tolerance}{tolerance for numerical noise in observation times. See \code{\link{num_leq_sorted_arrays}}.}

\item{\dots}{further arguments passed to or from methods.}
}
\value{
An vector of sampled values with same length as the input \code{x}.
}
\description{
Sample values from a time series at given sampling times.
}
\section{Methods (by class)}{
\itemize{
\item \code{uts}: sample from a \code{"uts"}
}}
\examples{
# Create a numeric "uts"

# Sample the most recent observation

# Sample with linear interpolatino
}

