% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_of_values.R
\name{sd}
\alias{sd}
\alias{sd.default}
\title{Generic sd function}
\usage{
sd(x, ...)

\method{sd}{default}(x, ...)
}
\arguments{
\item{x}{an \R object.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
The function is needed, because \code{\link[stats:sd]{sd}} of base \R is not generic.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: simply calls the default implementation of base \R
}}
\note{
As recommended in Section 7.1 ("Adding new generics") of "Writing R Extensions", the implementation of \code{\link{sd.default}} has been made a wrapper around \code{\link[stats:sd]{stats::sd}}.
}
\keyword{internal}

