% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/C_interfaces.R
\name{sorted_array_position}
\alias{sorted_array_position}
\title{Element Positions of a Sorted Array within Another Sorted Array}
\usage{
sorted_array_position(a, b, eps = 1e-09)
}
\arguments{
\item{a}{a sorted vector of numbers.}

\item{b}{a sorted vector of numbers.}

\item{eps}{tolerance for numerical noise, relative to the largest absolute elements in \code{a} and \code{b}}
}
\value{
A vector of indices of same length as vector \code{a}.
}
\description{
Assume given two sorted one-dimensional arrays \code{a} and \code{b}. For each element \code{a[i]}, determine the maximum index \code{j} in \code{b} such that \code{b[j] <= a[i]}, (\code{NA} if not found).
}
\examples{
sorted_array_position(1:5, 1:5)
sorted_array_position(c(-3, 1, 3, 5, 7), c(0, 1, 4, 9, 16))

# Numerical noise < eps has no effect
sorted_array_position(1, c(0, 1))
sorted_array_position(1 - 1e-13, c(0, 1))
sorted_array_position(1 + 1e-13, c(0, 1))
}

