% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/C_interfaces.R
\name{sorted_union}
\alias{sorted_union}
\title{Sorted Union}
\usage{
sorted_union(a, b, tolerance = 0)
}
\arguments{
\item{a}{a sorted vector of numbers.}

\item{b}{a sorted vector of numbers.}

\item{tolerance}{a non-negative number, indicating the tolerance for numerical noise. Specifically, numbers in \code{a} and \code{b} are appended one by one and in the appropriate order to the output vector. However, a number is only added if it is more than \code{tolerance} larger than the most recently added number.}
}
\value{
A numeric vector.
}
\description{
For two sorted numeric vectors \code{a} and \code{b}, determine the sorted union of elements.
}
\examples{
sorted_union(1:3, 2:4)

# Numerical noise < tolerance has no effect
sorted_union(0, 1e-14)
sorted_union(0, 1e-14, tolerance=1e-12)
sorted_union(c(0, 1e-14), 2, tolerance=1e-12)

# Trivial cases
sorted_union(1:10, c())
sorted_union(c(), 1:10)
}
\keyword{internal}

