% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_interfaces.R
\name{sorted_union_R}
\alias{sorted_union_R}
\title{R implementation of sorted_union}
\usage{
sorted_union_R(a, b)
}
\arguments{
\item{a}{a sorted vector of numbers.}

\item{b}{a sorted vector of numbers.}
}
\value{
A numeric vector.
}
\description{
This function is identical to \code{\link{sorted_union}} except that a) the input vectors don't need to be sorted, and b) there is no numerical noise tolerance support. It exists solely for testing the C implementation.
}
\examples{
sorted_union_R(1:3, 2:4)
}
\keyword{internal}
