% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{split_segments}
\alias{split_segments}
\title{Split Into Segments}
\usage{
split_segments(x, max_dt)
}
\arguments{
\item{x}{a \code{"uts"} object.}

\item{max_dt}{a non-negative \code{\link[lubridate]{duration}} object, specifying the maximum temporal spacing of consecutive observation times in the same segment.}
}
\value{
A list of \code{"uts"} objects. Consecutive observation times within each time series are less than \code{max_dt} apart.
}
\description{
Split a time series into the minimum number of segments (i.e. subperiod time series), such that consecutive observation times in the same segment are less than a specified value apart.
}
\examples{
split_segments(ex_uts(), ddays(0.25))
split_segments(ex_uts(), ddays(0.5))
split_segments(ex_uts(), ddays(Inf))
split_segments(ex_uts(), ddays(0))
}
\seealso{
\code{\link{plot}}, which optionally uses this function to not connect consecutive observations by a line.
}
\keyword{internal}

