% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uts.R
\name{start.uts}
\alias{end.uts}
\alias{start.uts}
\title{First and Last Observation Time}
\usage{
\method{start}{uts}(x, ...)

\method{end}{uts}(x, ...)
}
\arguments{
\item{x}{a \code{"uts"} object.}

\item{\dots}{further arguments passed to or from methods.}
}
\value{
\code{start()} returns the first observation time.

\code{end()} returns the last observation time.
}
\description{
Get the first and last observation time, respectively.
}
\examples{
start(ex_uts())
end(ex_uts())
}

