% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uts.R
\name{str.uts}
\alias{str.uts}
\title{Internal Structure of uts}
\usage{
\method{str}{uts}(object, ...)
}
\arguments{
\item{object}{a \code{"uts"} object.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Compactly display the internal structure of a \code{"uts"} object.
}
\note{
This methods only exists because the default implementation of \code{\link{str}} in base \R produces an error message.
}
\examples{
str(ex_uts())
str(ex_uts2())
}
\keyword{internal}

