% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uts.R
\name{summary.uts}
\alias{summary.uts}
\title{Object Summary}
\usage{
\method{summary}{uts}(object, ...)
}
\arguments{
\item{object}{a time series object.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Call \code{\link{summary}} on the values of a time series.
}
\note{
This method only exists because \code{\link{summary.default}} produces an error message when called with a \code{"uts"} object.
}
\examples{
summary(ex_uts())
}
\keyword{internal}

