% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_t_tail_t.R
\name{tail_t}
\alias{tail_t}
\alias{tail_t.uts}
\title{Terminal Subperiod}
\usage{
tail_t(x, ...)

\method{tail_t}{uts}(x, width, ...)
}
\arguments{
\item{x}{a time series object.}

\item{\dots}{further arguments passed to or from methods.}

\item{width}{a \code{\link[lubridate]{duration}} object, specifying the temporal width of the terminal subperiod time series.}
}
\description{
Return a subperiod time series with the terminal part, where the length is specified in terms of the \emph{temporal length}.
}
\section{Methods (by class)}{
\itemize{
\item \code{uts}: terminal subperiod time series for \code{"uts"} object.
}}

\note{
For an evenly spaced time series, the methods \code{tail()} and \code{tail_t()} essentially work the same. For unevenly spaced time series, however, they are quite different. The subperiod window of the former is determined by the observation \emph{values}, while for the latter it is determined by the observation \emph{times}.
}
\examples{
tail_t(ex_uts(), ddays(1))
tail_t(ex_uts(), ddays(0))  # leaves only the last observation
}
\seealso{
\code{\link{head}}, \code{\link{head_t}}, \code{\link{tail}}, \code{\link{window}} for other methods that extract a subperiod time series.
}
