% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uts.R
\name{time.uts}
\alias{time.uts}
\title{Observation Times}
\usage{
\method{time}{uts}(x, ...)
}
\arguments{
\item{x}{a \code{"uts"} object with \code{\link{logical}} observation values.}

\item{\dots}{further arguments passed to or from methods.}
}
\value{
A \code{\link{POSIXct}} object.
}
\description{
Get the observation times.
}
\note{
The observation times of a \code{"uts"} object \code{x} can also be accessed using \code{x$times}. However, using \code{time(x)} relies on the internal object representation.
}
\examples{
time(ex_uts())
identical(time(ex_uts()), ex_uts()$times)     # TRUE
}

