% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoindependent.R
\name{Two_Independent_Population}
\alias{Two_Independent_Population}
\alias{twoindependent}
\title{Two Independent Population}
\usage{
twoindependent(xbar, s, n, a = 0.05, mu = 0)
}
\arguments{
\item{xbar}{dadekno siji px2}

\item{s}{dadekno list(s1,s2)}

\item{n}{c(n1,n2)}
}
\description{
Two Independent Population
}
\author{
Muhammad Sirojuddin
}
