% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{uts-internal}
\alias{uts-internal}
\title{Internal Functions}
\description{
The internal functions listed below might be of interest to developers seeking to extend the package functionality.
}
\details{
\code{uts} methods:
\itemize{
  \item \code{\link{is.uts}}
  \item \code{\link{mean}}
  \item \code{\link{median}}
  \item \code{\link{sd}}
  \item \code{\link{str}}
  \item \code{\link{summary}}
}

Helper functions:
\itemize{
  \item \code{\link{num_leq_sorted}}
  \item \code{\link{num_leq_sorted_R}}
  \item \code{\link{sorted_union}}
  \item \code{\link{sorted_union_R}}
  \item \code{\link{split_segments}}
}

Functions in base \R that are not generic, but really ought to be:
\itemize{
  \item \code{\link{which}}
  \item \code{\link{which.default}}
  \item \code{\link{which.max}}
  \item \code{\link{which.max.default}}
  \item \code{\link{which.min}}
  \item \code{\link{which.min.default}}
}

Methods that exists solely ensure that methods intended for \code{\link{ts}} objects in base \R are not accidentally applied to \code{"uts"} objects:
\itemize{
  \item \code{\link{cycle}}
  \item \code{\link{frequency}}
}
}

