% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uts.R
\name{uts}
\alias{uts}
\title{Unevenly-spaced Time Series}
\usage{
uts(values = numeric(), times = as.POSIXct(character()))
}
\arguments{
\item{values}{a vector of observation values.}

\item{times}{a vector of strictly increasing observation times. Must be a \code{\link{POSIXct}} object or be coercible using \code{\link{as.POSIXct}}. Observation times cannot be \code{NA}.}
}
\value{
An object of class \code{"uts"}.
}
\description{
Create an unevenly spaced time series (\code{"uts"}) object from a vector of observation values and a vector of observation times of matching length.
}
\details{
As shown in the example below, it is possible to store arbitrary \R objects in a \code{"uts"} object. However, some time series operations (e.g. arithmetic operations) require the observation values to be either \code{\link{numeric}}, \code{\link{logical}}, or \code{\link{complex}}.
}
\note{
A virtual class \code{"uts_virtual"} exists from which \code{"uts"}, \code{"uts_vector"}, \code{"uts_matrix"}, and \code{"uts_data_frame"} inherit (see package \code{utsMultivariate} for the latter three classes): it is used to allow operations such as subtraction to mix the classes.
}
\examples{
# Create a numeric "uts"
dts <- c("2007-11-08", "2007-11-08", "2007-11-08", "2007-11-09", "2007-11-09", "2007-11-09")
tms <- c("7:00:00", "8:01:00", "13:15:00", "7:30:00", "8:51:00", "15:15:00")
uts(values=c(48.375, 48.5, 48.375, 47, 47.5, 47.35), times=paste(dts, tms))

# Create a "uts" with non-numeric observations
uts(list(1:5, c("a", "B")), c("2007-11-08 1:01:00", "2007-11-09 15:16:00"))

# Create an empty "uts"
uts()
}
\keyword{classes}
\keyword{ts}

