% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/which.R
\name{which.max}
\alias{which.max}
\alias{which.max.default}
\title{Generic which.max function}
\usage{
which.max(x, ...)

\method{which.max}{default}(x, ...)
}
\arguments{
\item{x}{an \R object.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
The function is needed, because \code{\link[base:which.max]{which.max}} of base \R is not generic.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: simply calls the default implementation of base \R
}}
\note{
As recommended in Section 7.1 ("Adding new generics") of "Writing R Extensions", the implementation of \code{\link{which.max.default}} has been made a wrapper around \code{\link[base:which.max]{base::which.max}}.
}
\keyword{internal}

