% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.R
\name{which.max.uts}
\alias{which.max.uts}
\alias{which.max.uts}
\alias{which.min.uts}
\title{Which Observation Time has the largest/smallest Observation Value?}
\usage{
\method{which.max}{uts}(x, ...)

\method{which.min}{uts}(x, ...)
}
\arguments{
\item{x}{a \code{"uts"} object.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Determine the temporal location (i.e. observation time) of the first maximum or minimum observation value.
}
\examples{
which.max(ex_uts())
which.min(ex_uts())
}
\seealso{
\code{\link[base]{which.max}} and \code{\link[base]{which.min}} for numeric vectors in base \R.
}
