% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uts.R
\name{which.uts}
\alias{which.uts}
\title{Which observation values are TRUE?}
\usage{
\method{which}{uts}(x, ...)
}
\arguments{
\item{x}{a \code{"uts"} object with \code{\link{logical}} observation values.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
For a logical \code{"uts"} (i.e. a \code{"uts"} with logical observation values), get the observation times with \code{TRUE} observation value.
}
\examples{
which(ex_uts() > 48)
}
\seealso{
\code{\link{which}}
}

