% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uts.R
\name{window.uts}
\alias{window.uts}
\title{Time Window}
\usage{
\method{window}{uts}(x, start = NULL, end = NULL, ...)
}
\arguments{
\item{x}{a \code{"uts"} object.}

\item{start, end}{\code{\link{POSIXct}} objects or coercible using \code{\link{as.POSIXct}}. The start and end times, respectively, of the period of interest.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Extract a subperiod time series between times \code{start} and \code{end}.
}
\examples{
window(ex_uts(), start="2007-11-09")
window(ex_uts(), start="2007-11-08", end="2007-11-09")
window(ex_uts(), start=as.POSIXct("2007-11-09", tz="America/New_York"))
window(ex_uts(), start=as.POSIXct("2007-11-09", tz="Australia/Sydney"))
}
\seealso{
\code{\link{head}}, \code{\link{head_t}}, \code{\link{tail}}, \code{\link{tail_t}} for other methods that extract a subperiod time series.
}
