% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uts.R
\name{as.data.frame.uts}
\alias{as.data.frame.uts}
\title{Coerce to a Data Frame}
\usage{
\method{as.data.frame}{uts}(x, ...)
}
\arguments{
\item{x}{a \code{"uts"} object.}

\item{\dots}{arguments passed to \code{\link{format.POSIXct}}.}
}
\description{
Flatten a \code{"uts"} to a two-column \code{data.frame}, with the observation times in first column and observation values in the second column. The column names are \code{"time"} and \code{"value"}.
}
\note{
Only time series with atomic observation values can be coerced to a \code{data.frame}.
}
\examples{
as.data.frame(ex_uts())
as.data.frame(uts(c("cat", "dog"), Sys.time() + days(1:2)), format="\%Y-\%m-\%d")
}

