% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coercions.R
\name{as.uts}
\alias{as.uts}
\alias{as.uts.irts}
\alias{as.uts.ts}
\alias{as.uts.xts}
\alias{as.uts.zoo}
\title{Coercion to uts}
\usage{
as.uts(x, ...)

\method{as.uts}{ts}(x, ...)

\method{as.uts}{xts}(x, ...)

\method{as.uts}{zoo}(x, ...)

\method{as.uts}{irts}(x, ...)
}
\arguments{
\item{x}{a time series object of appropriate type.}

\item{\dots}{further arguments passed to or from methods.}
}
\value{
An object of class \code{"uts"}.
}
\description{
Convert other time series objects to \code{"uts"} objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{ts}: convert a \code{\link[stats:ts]{ts}} object

\item \code{xts}: convert a \code{\link[xts:xts]{xts}} object

\item \code{zoo}: convert a \code{\link[zoo:zoo]{zoo}} object

\item \code{irts}: convert a \code{\link[tseries:irts]{irts}} object
}}
\examples{
# Convert a quarterly "ts" time series
ts1 <- ts(1:10, frequency = 4, start = c(1959, 2))
as.uts(ts1)

# Convert a monthly "ts" time series
ts2 <- ts(1:10, frequency = 12, start = c(1959, 8))
as.uts(ts2)

# Convert a yearly 'ts" time series
ts3 <- ts(1:10, frequency = 1, start = 1959)
as.uts(ts3)
#
# Convert an "xts" time series
if (requireNamespace("xts", quietly = TRUE)) {
  xts1 <- xts::xts(1:4, as.Date("2015-01-01") + c(1, 3, 7, 9))
  as.uts(xts1)
}
#
# Convert a "zoo" time series
if (requireNamespace("zoo", quietly = TRUE)) {
  zoo1 <- zoo::zoo(1:4, as.Date("2015-01-01") + c(1, 3, 7, 9))
  as.uts(zoo1)
}
#
# Convert an "irts" time series
if (requireNamespace("tseries", quietly = TRUE)) {
  irts1 <- tseries::irts(as.POSIXct("2015-01-01") + days(c(1, 3, 7, 9)), 1:4)
  as.uts(irts1)
  
  # Multivariate 'irts' objects need to be converted using as.uts_vector()
  \dontrun{
     t <- cumsum(rexp(10, rate = 0.1))
     v <- matrix(rnorm(20), nrow=10)
     irts2 <- tseries::irts(t, v)
     as.uts(irts2)
  }
}
}

