% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_compatability.R
\name{compatability}
\alias{as.ts.uts}
\alias{compatability}
\alias{cycle.uts}
\alias{frequency.uts}
\title{Compatability with ts class}
\usage{
\method{as.ts}{uts}(x, ...)

\method{cycle}{uts}(x, ...)

\method{frequency}{uts}(x, ...)
}
\arguments{
\item{x}{a \code{"uts"} object.}

\item{\dots}{further arguments passed to or from methods.}
}
\value{
\code{cycle()} and \code{frequency()} give an error message, because \code{"uts"} objects, by definition, do not have a fixed number of observations in a given time interval.
}
\description{
These methods exist solely to ensure that methods intended for \code{"ts"} objects in base \R are not accidentally applied to \code{"uts"} objects.
}
\examples{
\dontrun{as.ts(ex_uts())}
\dontrun{cycle(ex_uts())}
\dontrun{frequency(ex_uts())}
}
\seealso{
\code{\link{ts}}
}
\keyword{internal}

