% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diff.R
\name{diff.uts}
\alias{diff.uts}
\title{Lagged Differences}
\usage{
\method{diff}{uts}(x, lag = 1, scale = "log", ...)
}
\arguments{
\item{x}{a \code{"uts"} object.}

\item{lag}{an integer indicating which lag to use.}

\item{scale}{on which scale to calculate lagged differences. Either \code{"abs"} for absolute differences \code{X[n] - X[n - lag]}, \code{"rel"} for relative differences \code{X[n] / X[n - lag] - 1}, or \code{"log"} for logarithmic differences \code{log(X[n] / X[n - lag])}.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Return a time series with suitably lagged differences between the observation values.
}
\note{
Observations times without corresponding lagged value (for example, the second observation for \code{lag=3}) are dropped from the output.
}
\examples{
diff(ex_uts())
diff(ex_uts(), lag=-3)
diff(ex_uts(), scale="log")
diff(ex_uts(), lag=10)     # an empty time series, because the lag is too large
}
\seealso{
\code{\link{lag}}, \code{\link{lag_t}}, and \code{\link{diff_t}}
}

