% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diff.R
\name{diff_t}
\alias{diff_t}
\title{Rolling Differences}
\usage{
diff_t(x, ...)
}
\arguments{
\item{x}{a time series object.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Rolling Differences
}

