% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_last.R
\name{first}
\alias{first}
\alias{first.uts}
\title{First Observation Value}
\usage{
first(x, ...)

\method{first}{uts}(x, ...)
}
\arguments{
\item{x}{an \R object.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Return the first value of an \R object.
}
\section{Methods (by class)}{
\itemize{
\item \code{uts}: Return the first observation \emph{value} of a \code{"uts"} object. Unlike method \code{first()} in package \href{https://cran.r-project.org/web/packages/xts/}{xts}, arguments \code{n} and \code{keep} are not supported. Instead, this functionality is provided by \code{head()} and \code{heat_t()}.
}}
\examples{
first(ex_uts())
}
\seealso{
\code{\link{last}}

\code{\link{start}} and \code{\link{end}} return, respectively, the first and last obvservation \emph{time} from a time series.
}

