% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/operators.R
\name{group_generic_uts}
\alias{Math.uts}
\alias{Summary.uts}
\alias{group_generic_uts}
\title{S3 Group Generic Methods for uts}
\usage{
\method{Summary}{uts}(x, ...)

\method{Math}{uts}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"uts"}.}

\item{\dots}{further arguments passed to or from methods.}
}
\value{
An object of class \code{"uts"} with same observation times as the input \code{x}, and observation values the result of applying the provided method to the observation values of input \code{x}.
}
\description{
These methods apply the methods in base \R{} to the observation values of \code{\link{uts}} objects.
}
\examples{
# "Summary" methods
min(ex_uts())
#any(ex_uts() > 45)
range(ex_uts())
#
# "Math" methods
floor(ex_uts())
ceiling(ex_uts())
sqrt(ex_uts())
cumsum(ex_uts())
}
\seealso{
\code{\link{groupGeneric}}
}

