% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_tail.R
\name{head.uts}
\alias{head.uts}
\alias{tail.uts}
\title{Return the First or Last Part}
\usage{
\method{head}{uts}(x, n = 6L, ...)

\method{tail}{uts}(x, n = 6L, ...)
}
\arguments{
\item{x}{a \code{"uts"} object.}

\item{n}{a single integer. If positive, length of the the resulting time series. If negative, all but the \code{n} last/first number of observations of \code{x}.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Return a subperiod time series with the first or last part. The length of the output is specified in terms of the \emph{number of observations}.
}
\note{
For an evenly spaced time series, the methods \code{head()} and \code{head_t()} (or \code{tail()} and \code{tail_t()}) essentially work the same. For unevenly spaced time series, however, they are quite different. The subperiod window of the former is determined by the observation \emph{values}, while for the latter it is determined by the observation \emph{times}.
}
\examples{
head(ex_uts(), 2)
head(ex_uts(), -2)
head(ex_uts(), -6)
tail(ex_uts(), 1)
tail(ex_uts(), -1)
tail(ex_uts(), -6)
}
\seealso{
\code{\link{head_t}}, \code{\link{tail_t}}, \code{\link{window}} for other methods that extract a subperiod time series.
}

