% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAs.R
\name{is.na.uts}
\alias{is.na.uts}
\title{Not Available / Missing Observation Values}
\usage{
\method{is.na}{uts}(x)
}
\arguments{
\item{x}{a \code{"uts"} object.}
}
\value{
A logical \code{"uts"} (i.e. a \code{"uts"} with \code{\link{logical}} observation values), indicating which observation values are \code{NA}.
}
\description{
Find missing values.
}
\examples{
# Set observation to NA
test <- ex_uts()
test$values[c(2, 4)] <- NA      # NEXT: use replacement using time points

# Get logical "uts", indicating which observations are NA
is.na(test)
}
\seealso{
\code{\link{na.omit.uts}}

\code{\link{is.na}}, \code{\link{na.omit}} in base \R.
}

