% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lag.R
\name{lag.uts}
\alias{lag.uts}
\title{Lag Obseration Values}
\usage{
\method{lag}{uts}(x, k = 1, ...)
}
\arguments{
\item{x}{a \code{"uts"} object.}

\item{k}{the number of lags (in units of observations).}

\item{\dots}{further arguments passed to or from methods.}
}
\value{
A \code{"uts"} object with the observation times (and in particular, of the same length) as \code{x}.
}
\description{
Compute a lagged version of a time series by shifting individual observations values, while keeping the observation times unchanged.
}
\details{
The n-th observation of each original time series becomes the (n+k)-th observation of the lagged time series for 1 <= (n+k) <= length(x). Observations without corresponding un-lagged value (for example, the second observation for lag \code{k=3}) are set to \code{NA}.
}
\note{
For an evenly-spaced time series (1) shifting observation \emph{times}, and (2) shifting observation \emph{values} essentinally gives the same result (apart from the \code{NA}s that are introduced in the latter case). For unevenly-spaced time series, however, these two operations are quite different. The former affects only the observation times (but not the observation values), while the latter affects only the observation values (but not the observation times).
}
\examples{
# Shift observations values forward by one observation
lag(ex_uts(), k=1)

# Shift observations values forward by two observations
lag(ex_uts(), k=-2)

# If the lag >= the length of the time series, all observation values are N
lag(ex_uts(), k=6)
lag(ex_uts(), k=-6)
}
\seealso{
\code{\link[stats:lag]{lag}} in base \R.

#'

\code{\link{lag_t}} allows to shift observation \emph{times}, as opposed to observation \emph{values}.
}

