% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag.R
\name{lag.uts}
\alias{lag.uts}
\title{Lag Observation Values}
\usage{
\method{lag}{uts}(x, k = 1, ...)
}
\arguments{
\item{x}{a \code{"uts"} object.}

\item{k}{the number of lags (in units of observations).}

\item{\dots}{further arguments passed to or from methods.}
}
\value{
A \code{"uts"} object with the same observation times (apart from dropped observations) as \code{x}.
}
\description{
Compute a lagged version of a time series by shifting observation values.
}
\details{
Each observation time-value tuple \eqn{(t[n], x[n])} in the original time series is replaced by \eqn{(t[n], x[n-k])} in the lagged time series. Observations without corresponding un-lagged value (for example, the second observation for lag \code{k=3}) are dropped from the output.
}
\note{
For an evenly spaced time series (1) shifting observation \emph{times}, and (2) shifting observation \emph{values} essentinally gives the same result. For unevenly spaced time series, however, these two operations are quite different. The former affects only the observation \emph{times}, while the latter affects only the observation \emph{values} (apart from observations that are dropped).
}
\examples{
# Shift observations values forward by one observation
lag(ex_uts(), k=1)

# Shift observations values forward by two observations
lag(ex_uts(), k=-2)

# If the lag 'k' is >= the length of the time series, all observations are dropped
lag(ex_uts(), k=6)
}
\seealso{
\code{\link{lag_t}} allows to shift observation \emph{times}, as opposed to observation \emph{values}.

\code{\link[stats:lag]{lag}} in base \R.
}

