% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/first_last.R
\name{last}
\alias{last}
\alias{last.uts}
\title{Last Value}
\usage{
last(x, ...)

\method{last}{uts}(x, ...)
}
\arguments{
\item{x}{an \R object.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Return the last value of an \R object.
}
\section{Methods (by class)}{
\itemize{
\item \code{uts}: Return the last observation \emph{value} of a \code{"uts"} object. Unlike method \code{last()} in package \href{https://cran.r-project.org/web/packages/xts/}{xts}, arguments \code{n} and \code{keep} are not supported. Instead, this functionality is provided by \code{tail()} and \code{tail_t()}.
}}
\examples{
first(ex_uts())
}
\seealso{
\code{\link{first}}

\code{\link{start}} and \code{\link{end}} return, respectively, the first and last obvservation \emph{time} from a time series.
}

