% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uts.R
\name{length.uts}
\alias{length.uts}
\title{Number of Observations}
\usage{
\method{length}{uts}(x)
}
\arguments{
\item{x}{a \code{"uts"} object.}
}
\description{
Return the number of time series observations.
}
\note{
This methods only exists because the default implementation of \code{\link{length}} from base \R returns the number of fields in the internal representation of a \code{"uts"} object, instead of the number of observations.
}
\examples{
length(ex_uts())
}
\keyword{internal}

