% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{length.uts}
\alias{length.uts}
\title{Number of Observations}
\usage{
\method{length}{uts}(x)
}
\arguments{
\item{x}{a \code{"uts"} object.}
}
\description{
Return the number of time series observations.
}
\note{
For an evenly spaced time series there is a one-to-one correspondence between (1) the \emph{number} of observations, and (2) \emph{temporal} length. For unevenly spaced time series, however, these two concepts are quite different.s.
}
\examples{
length(ex_uts())
length(uts())
}
\seealso{
\code{\link{length_t}} allows to get the temporal length of a time series.
}
