% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary_of_values.R
\name{mean.uts}
\alias{mean.uts}
\alias{sd.uts}
\title{Mean and Standard Deviation}
\usage{
\method{mean}{uts}(x, ...)

\method{sd}{uts}(x, ...)
}
\arguments{
\item{x}{a \code{"uts"} object.}

\item{\dots}{further arguments passed to or from methods.}
}
\value{
\code{mean()} returns the arithmetic mean of the observation values.

\code{sd()} returns the standard devation of t he observation values.
}
\description{
Calculate the mean and standard deviation of the observation values.
}
\examples{
mean(ex_uts())
sd(ex_uts())
}
\seealso{
\code{\link[base:mean]{mean}} and \code{\link[stats:sd]{sd}} in base \R.
}

