% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_of_values.R
\name{mean.uts}
\alias{mean.uts}
\alias{median.uts}
\alias{sd.uts}
\title{Mean, Median, and Standard Deviation}
\usage{
\method{mean}{uts}(x, ...)

\method{median}{uts}(x, ...)

\method{sd}{uts}(x, ...)
}
\arguments{
\item{x}{a \code{"uts"} object.}

\item{\dots}{further arguments passed to or from methods.}
}
\value{
\code{mean()} returns the arithmetic mean of the observation values.

\code{median()} returns the median of the observation values.

\code{sd()} returns the standard devation of the observation values.
}
\description{
Calculate the mean, median, and standard deviation, respectively, of the observation values.
}
\examples{
mean(ex_uts())
median(ex_uts())
sd(ex_uts())
}
\seealso{
\code{\link[base:mean]{mean}}, \code{\link[stats:median]{median}}, and \code{\link[stats:sd]{sd}} in base \R.
}

