% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uts.R
\name{merge.uts}
\alias{merge.uts}
\title{Merge two or more uts}
\usage{
\method{merge}{uts}(x, y, tolerance = .Machine$double.eps^0.5, ...)
}
\arguments{
\item{x, y}{\code{"uts"} objects.}

\item{tolerance}{tolerance for numerical noise in observation times.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Merge two or more \code{"uts"} into a single time series. For observation times that show up in more than one time series, the observation value of the first \code{"uts"} in the argument list with such observation time is used.
}
\examples{
merge(ex_uts(), ex_uts())
merge(ex_uts(), ex_uts2())
merge(ex_uts2(), ex_uts())
merge(uts(), ex_uts())
}
