% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAs.R
\name{na.omit.uts}
\alias{na.omit.uts}
\title{Remove NA Observation Values}
\usage{
\method{na.omit}{uts}(object, ...)
}
\arguments{
\item{object}{a time series object.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Returns the object with incomplete cases removed.
}
\examples{
# Remove NAs from a "uts"
test <- ex_uts()
test$values[c(2, 4)] <- NA
na.omit(test)
}
\seealso{
\code{\link{is.na.uts}}

\code{\link{is.na}}, \code{\link{na.omit}} in base \R.
}

