% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uts.R
\name{na.omit.uts}
\alias{na.omit.uts}
\title{Remove NAs}
\usage{
\method{na.omit}{uts}(object, ...)
}
\arguments{
\item{object}{a time series object.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Returns the object with incomplete cases removed.
}
\examples{
# Remove NAs from a "uts"
tmp <- ex_uts()
tmp$values[c(2, 4)] <- NA
na.omit(tmp)
}
\seealso{
\code{\link[stats:na.fail]{na.fail}}, \code{\link[stats:na.fail]{na.omit}}
}

