% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/POSIXt_extensions.R
\name{next_business_day.POSIXt}
\alias{next_business_day.POSIXt}
\title{Next Business Day}
\usage{
\method{next_business_day}{POSIXt}(x, ...)
}
\arguments{
\item{x}{a \code{POSIXt} date-time object.}

\item{\dots}{further arguments passed to or from methods.}
}
\value{
An object of the same class and length as the input \code{x}.
}
\description{
For a POSIXt date-time object, add the minimum number of days to reach a business day. The time-of-day information is therefore unaffected. In particular, if the input already is a business day, then the output is the same as the input.
}
\note{
Non-business days are defined as Saturdays and Sundays. Public holidays are not supported.
}
\examples{
# 2015-04-20 is a Monday
next_business_day(as.POSIXct("2015-04-20") + days(0:6))
}
\seealso{
\code{\link{previous_business_day.POSIXt}}
}
\keyword{chron}

