% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_methods.R
\name{Ops.uts_virtual}
\alias{Ops.uts}
\alias{Ops.uts_virtual}
\title{Ops Group Methods}
\usage{
\method{Ops}{uts_virtual}(e1, e2)
}
\arguments{
\item{e1, e2}{either \code{"uts"} objects or compatible \R objects of length one, where compatability depends on the type of operation performed.}
}
\description{
Apply the \code{\link{Ops}} group methods in base \R{} to the observation values of \code{"uts"} objects.
}
\details{
For compatability with the S3 classes defined in the \code{utsMultivariate} package, the Ops group methods are implemented via \code{Ops.list} instead of \code{Ops.uts}.
}
\note{
For unary oparations, the output time series has the same observation times as the input time series.

For binary operations involving two time series \code{e1} and \code{e2}, the output time series has the union of observation times of \code{e1} and \code{e2}, but excluding times before the first observation time of either time series. The method for determining these times is unaffected by numerical noise less than \code{sqrt(\link[=.Machine]{.Machine$double.eps})}.
}
\examples{
# Unary oparators
-ex_uts()
!ex_uts()

# Binary operators
ex_uts() * 2
2 * ex_uts()
ex_uts() / ex_uts()
ex_uts() > 48
48 >= ex_uts()
}
\seealso{
\code{\link{groupGeneric}}
}

