% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.uts}
\alias{plot.uts}
\title{Plot a uts}
\usage{
\method{plot}{uts}(x, max_dt = ddays(Inf), type = "l", col = "blue",
  xlab = "", ylab = "", ...)
}
\arguments{
\item{x}{a \code{"uts"} object with numeric or logical observation values.}

\item{max_dt}{a non-negative \code{\link[lubridate]{duration}} object. Consecutive observations are not connected by a line in the graph, if they are more than this amount apart in time.}

\item{type}{what type of plot should be drawn, see \code{\link[graphics:plot.default]{plot.default}}. \code{type="b"} is helpful for highlighting individual observations.}

\item{col}{the colors for lines and points. See \code{\link[graphics:plot.default]{plot.default}}.}

\item{xlab}{a label for the x axis.}

\item{ylab}{a label for the y axis.}

\item{\dots}{other graphical parameters passed to \code{\link[graphics:plot.default]{plot.default}}.}
}
\description{
A convenience wrapper around \code{\link[graphics:plot.default]{plot.default}} with several sensible default arguments.
}
\examples{
plot(ex_uts())
plot(ex_uts(), max_dt=dhours(12), type="b")   # don't connect points more than 12 hours apart
plot(ex_uts(), max_dt=dhours(2), type="b")

# Plot time series with NAs
tmp <- ex_uts()
tmp$values[2] <- NA
plot(tmp, type="o")         # isolated points are plotted as circles
plot(na.omit(tmp), type="b")

# Use custom date formatting for x axis
# -) the "format" argument is passed down to axis.POSIXct()
# -) this example produces several harmless warning messages, because other
#    functions that are part of the call chain have no "format" argument
plot(ex_uts(), format="\%a \%I\%p")
}

