% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/POSIXct_matrix.R
\name{POSIXct_matrix}
\alias{POSIXct_matrix}
\title{Matrix of POSIXct Objects}
\usage{
POSIXct_matrix(x, ...)
}
\arguments{
\item{x}{an \R{} object of appropriate type.}

\item{\dots}{arguments passed to or from methods.}
}
\description{
Generic function to create a matrix of POSIXct Objects.
}
\details{
See documentation on method functions for further details.
}
\seealso{
\code{\link{POSIXct_vector}}
}
\keyword{chron}
\keyword{classes}

