% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/POSIXt_extensions.R
\name{quarterend}
\alias{quarterend}
\title{Quarter End}
\usage{
quarterend(x, ...)
}
\arguments{
\item{x}{an object representing dates.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Generic function to determine the quarter-end of objects representing dates.
}
\details{
See documentation on method functions for further details.
}
\seealso{
\code{\link[lubridate:ceiling_date]{ceiling_date()}}
}

