% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_values}
\alias{sample_values}
\alias{sample_values.uts}
\title{Sample Values}
\usage{
sample_values(x, ...)

\method{sample_values}{uts}(x, time_points, interpolation = "last",
  max_dt = ddays(Inf), tolerance = .Machine$double.eps^0.5, ...)
}
\arguments{
\item{x}{a time series object.}

\item{time_points}{a strictly increasing sequence of \code{\link{POSIXct}} date-times.}

\item{interpolation}{either \code{"last"} or \code{"linear"}, specifying the interpolation method.
\code{"last"} returns the most recent observation value in \code{x} for each sampling time.
\code{"linear"} returns for each sampling time the linearily-interpolated  observation values in \code{x} that immediately preceede and follow each sampling time.}

\item{max_dt}{a duration object, specifying the maximum time difference between each sampling time and the observation times in \code{x} used for determining the sampled value. Return \code{NA} as sampled value for sampling times for which this threshold is exceeded.}

\item{tolerance}{tolerance for numerical noise in observation times. See \code{\link{num_leq_sorted}}.}

\item{\dots}{further arguments passed to or from methods.}
}
\value{
An vector of sampled values with same length as the input \code{x}.
}
\description{
Sample observation values from a time series at given sampling times.
}
\section{Methods (by class)}{
\itemize{
\item \code{uts}: sample from a \code{"uts"} object
}}
\examples{
# Sample the most recent observation
times <- as.POSIXct(c("2007-11-09", "2007-11-10"))
sample_values(ex_uts(), times)

# Sample with linear interpolation
sample_values(ex_uts(), times, interpolation="linear")

# Sample a non-numeric time series
sample_values(ex_uts2(), times)

# Error, because only numeric time series can be linearly interpolated
\dontrun{sample_values(ex_uts2(), as.POSIXct("2007-01-01"), interpolation="linear")}
}

