% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampling.R
\name{[.uts}
\alias{[.uts}
\title{Extract or Replace Parts of a uts}
\usage{
\method{[}{uts}(x, time_points, ...)
}
\arguments{
\item{x}{a \code{"uts"} object.}

\item{time_points}{either a strictly increasing sequence of \code{\link{POSIXct}} date-times, or a \code{uts} with \code{\link{logical}} observation values.}

\item{\dots}{further arguments passed to \code{\link{sample_values}}.}
}
\description{
The accessor method extracts a sub-sampled time series with the provided times. The replacement method inserts new observation values at the provided observation times, replacing observations values for already existing observation times (if any).
}
\examples{
ex_uts()[as.POSIXct("2000-01-01")]
#ex_uts()[alltimes(utsv, 2)]
#
times <- as.POSIXct(c("2007-11-08 1:01:00", "2007-11-09  15:16:00"))
ex_uts()[times, max_lag = ddays(1)]
#ex_uts()[!is.na(ex_uts())]
#ex_uts()[ex_uts() > 48]
}

