% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_of_values.R
\name{summary.uts}
\alias{summary.uts}
\title{Summary of Time Series Values}
\usage{
\method{summary}{uts}(object, ...)
}
\arguments{
\item{x}{a \code{"uts"} object.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Apply \code{\link{summary}} from base \R to the observation values of a time series.
}
\note{
This method only exists because \code{\link{summary.default}} produces an error message.
}
\examples{
summary(ex_uts())
}
\keyword{internal}

