% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/operators.R
\name{Summary.uts}
\alias{Summary.uts}
\title{Summary Group Methods for uts}
\usage{
\method{Summary}{uts}(x, ...)
}
\arguments{
\item{x}{a \code{"uts"} object.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Apply the \code{\link{Summary}} methods in base \R{} to the observation values of \code{"uts"} objects.
}
\examples{
# Get the smallest observation value, ignoring NAs
min(ex_uts(), na.rm=TRUE)

# Check if any observation value is larger than 48
#any(ex_uts() > 48)
}
\seealso{
\code{\link{groupGeneric}}
}

