% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uts.R
\name{uts}
\alias{uts}
\title{Unevenly-spaced Time Series}
\usage{
uts(values = c(), times = as.POSIXct(character(0)))
}
\arguments{
\item{values}{a vector of observation values.}

\item{times}{a vector of observation times. Must be of class \code{"POSIXct"} or be coercible using \code{\link{as.POSIXct}}.}
}
\value{
An object of class \code{"uts"}.
}
\description{
Create an unevenly-spaced time series (\code{"uts"}) object from a vector of observation values and a vector of observation times of matching length.
}
\examples{
# Create a numeric "uts"
dts <- c("2007-11-08", "2007-11-08", "2007-11-08", "2007-11-09", "2007-11-09", "2007-11-09")
tms <- c("7:00:00", "8:01:00", "13:15:00", "7:30:00", "8:51:00", "15:15:00")
uts(values=c(48.375, 48.5, 48.375, 47, 47.5, 47.35), times=paste(dts, tms))

# Create an "uts" with non-numeric observations
uts(list(1:5, c("a", "B")), c("2007-11-08 1:01:00", "2007-11-09 15:16:00"))

# Create an empty "uts"
uts()
}
\keyword{classes}
\keyword{ts}

