% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uts.R
\name{window.uts}
\alias{window.uts}
\title{Time Window}
\usage{
\method{window}{uts}(x, start = NULL, end = NULL, ...)
}
\arguments{
\item{x}{a \code{"uts"} object.}

\item{start}{the start time of the period of interest.}

\item{end}{the end time of the period of interest.}

\item{\dots}{further arguments passed to or from methods.}
}
\description{
Extract a subperiod time series between times \code{start} and \code{end}.
}
\examples{
window(ex_uts(), start=as.POSIXct("2007-11-09"))
window(ex_uts(), start=as.POSIXct("2007-11-08"), end=as.POSIXct("2007-11-09"))
}

